refactor: split engines into lib files
