+++ src/do.ts
@@ -1,1085 +1,10 @@
 import { DurableObject } from "cloudflare:workers";
 
-// Types for better extensibility
-interface TableSchema {
-	columns: Record<string, string>;
-	sample_data: any[];
-	relationships?: Record<string, RelationshipInfo>;
-}
+import { SchemaInferenceEngine } from "./lib/SchemaInferenceEngine.js";
+import { DataInsertionEngine } from "./lib/DataInsertionEngine.js";
+import { PaginationAnalyzer } from "./lib/PaginationAnalyzer.js";
+import { TableSchema, ProcessingResult, PaginationInfo } from "./lib/types.js";
 
-interface RelationshipInfo {
-	type: 'foreign_key' | 'junction_table';
-	target_table: string;
-	foreign_key_column?: string;
-	junction_table_name?: string;
-}
-
-interface ProcessingResult {
-	success: boolean;
-	message?: string;
-	schemas?: Record<string, SchemaInfo>;
-	table_count?: number;
-	total_rows?: number;
-	pagination?: PaginationInfo;
-}
-
-interface SchemaInfo {
-	columns: Record<string, string>;
-	row_count: number;
-	sample_data: any[];
-	relationships?: Record<string, RelationshipInfo>;
-}
-
-interface PaginationInfo {
-	hasNextPage: boolean;
-	hasPreviousPage: boolean;
-	currentCount: number;
-	totalCount: number | null;
-	endCursor: string | null;
-	startCursor: string | null;
-	suggestion?: string;
-}
-
-interface EntityContext {
-	entityData?: any;
-	parentTable?: string;
-	parentKey?: string;
-	relationshipType?: 'one_to_one' | 'one_to_many' | 'many_to_many';
-}
-
-// Enhanced schema inference engine with proper relational decomposition
-class SchemaInferenceEngine {
-	private discoveredEntities: Map<string, any[]> = new Map();
-	private entityRelationships: Map<string, Set<string>> = new Map(); // Now tracks unique relationships only
-	
-	inferFromJSON(data: any): Record<string, TableSchema> {
-		// Reset state for new inference
-		this.discoveredEntities.clear();
-		this.entityRelationships.clear();
-		
-		const schemas: Record<string, TableSchema> = {};
-		
-		this.discoverEntities(data, []);
-		
-		// Only proceed if we found meaningful entities
-		if (this.discoveredEntities.size > 0) {
-			this.createSchemasFromEntities(schemas);
-		} else {
-			// Fallback for simple data
-			if (typeof data !== 'object' || data === null || Array.isArray(data)) {
-				const tableName = Array.isArray(data) ? 'array_data' : 'scalar_data';
-				schemas[tableName] = this.createSchemaFromPrimitiveOrSimpleArray(data, tableName);
-			} else {
-				schemas.root_object = this.createSchemaFromObject(data, 'root_object');
-			}
-		}
-
-		return schemas;
-	}
-	
-	private discoverEntities(obj: any, path: string[], parentEntityType?: string): void {
-		if (!obj || typeof obj !== 'object') {
-			return;
-		}
-
-		if (Array.isArray(obj)) {
-			if (obj.length > 0) {
-				// Process all items in the array - they should be the same entity type
-				let arrayEntityType: string | null = null;
-				
-				for (const item of obj) {
-					if (this.isEntity(item)) {
-						if (!arrayEntityType) {
-							arrayEntityType = this.inferEntityType(item, path);
-						}
-						
-						// Add to discovered entities
-						const entitiesOfType = this.discoveredEntities.get(arrayEntityType) || [];
-						entitiesOfType.push(item);
-						this.discoveredEntities.set(arrayEntityType, entitiesOfType);
-						
-						// Record relationship if this array belongs to a parent entity
-						if (parentEntityType && path.length > 0) {
-							const fieldName = path[path.length - 1];
-							if (fieldName !== 'nodes' && fieldName !== 'edges') { // Skip GraphQL wrapper fields
-								this.recordRelationship(parentEntityType, arrayEntityType);
-							}
-						}
-						
-						// Recursively process nested objects within this entity
-						this.processEntityProperties(item, arrayEntityType);
-					}
-				}
-			}
-			return;
-		}
-
-		// Handle GraphQL edges pattern
-		if (obj.edges && Array.isArray(obj.edges)) {
-			const nodes = obj.edges.map((edge: any) => edge.node).filter(Boolean);
-			if (nodes.length > 0) {
-				this.discoverEntities(nodes, path, parentEntityType);
-			}
-			return;
-		}
-
-		// Process individual entities
-		if (this.isEntity(obj)) {
-			const entityType = this.inferEntityType(obj, path);
-			
-			// Add to discovered entities
-			const entitiesOfType = this.discoveredEntities.get(entityType) || [];
-			entitiesOfType.push(obj);
-			this.discoveredEntities.set(entityType, entitiesOfType);
-			
-			// Process properties of this entity
-			this.processEntityProperties(obj, entityType);
-			return;
-		}
-
-		// For non-entity objects, recursively explore their properties
-		for (const [key, value] of Object.entries(obj)) {
-			this.discoverEntities(value, [...path, key], parentEntityType);
-		}
-	}
-	
-	private processEntityProperties(entity: any, entityType: string): void {
-		for (const [key, value] of Object.entries(entity)) {
-			if (Array.isArray(value) && value.length > 0) {
-				// Check if this array contains entities
-				const firstItem = value.find(item => this.isEntity(item));
-				if (firstItem) {
-					const relatedEntityType = this.inferEntityType(firstItem, [key]);
-					this.recordRelationship(entityType, relatedEntityType);
-					
-					// Process all entities in this array
-					value.forEach(item => {
-						if (this.isEntity(item)) {
-							const entitiesOfType = this.discoveredEntities.get(relatedEntityType) || [];
-							entitiesOfType.push(item);
-							this.discoveredEntities.set(relatedEntityType, entitiesOfType);
-							
-							// Recursively process nested entities
-							this.processEntityProperties(item, relatedEntityType);
-						}
-					});
-				}
-			} else if (value && typeof value === 'object' && this.isEntity(value)) {
-				// Single related entity
-				const relatedEntityType = this.inferEntityType(value, [key]);
-				this.recordRelationship(entityType, relatedEntityType);
-				
-				const entitiesOfType = this.discoveredEntities.get(relatedEntityType) || [];
-				entitiesOfType.push(value);
-				this.discoveredEntities.set(relatedEntityType, entitiesOfType);
-				
-				// Recursively process nested entities
-				this.processEntityProperties(value, relatedEntityType);
-			}
-		}
-	}
-	
-	private isEntity(obj: any): boolean {
-		if (!obj || typeof obj !== 'object' || Array.isArray(obj)) return false;
-		
-		// An entity typically has an ID field or multiple meaningful fields
-		const hasId = obj.id !== undefined || obj._id !== undefined;
-		const fieldCount = Object.keys(obj).length;
-		const hasMultipleFields = fieldCount >= 2;
-		
-		// Check for common entity patterns
-		const hasEntityFields = obj.name !== undefined || obj.title !== undefined || 
-			obj.description !== undefined || obj.type !== undefined;
-		
-		return hasId || (hasMultipleFields && hasEntityFields);
-	}
-	
-	private inferEntityType(obj: any, path: string[]): string {
-		// Try to infer type from object properties (e.g., __typename)
-		if (obj.__typename) return this.sanitizeTableName(obj.__typename);
-		if (obj.type && typeof obj.type === 'string' && !['edges', 'node'].includes(obj.type.toLowerCase())) {
-			return this.sanitizeTableName(obj.type);
-		}
-		
-		// Infer from path context, attempting to singularize
-		if (path.length > 0) {
-			let lastName = path[path.length - 1];
-
-			// Handle GraphQL patterns
-			if (lastName === 'node' && path.length > 1) {
-				lastName = path[path.length - 2];
-				if (lastName === 'edges' && path.length > 2) {
-					lastName = path[path.length - 3];
-				}
-			} else if (lastName === 'edges' && path.length > 1) {
-				lastName = path[path.length - 2];
-			}
-			
-			// Attempt to singularize common plural forms
-			const sanitized = this.sanitizeTableName(lastName);
-			if (sanitized.endsWith('ies')) {
-				return sanitized.slice(0, -3) + 'y';
-			} else if (sanitized.endsWith('s') && !sanitized.endsWith('ss') && sanitized.length > 1) {
-				const potentialSingular = sanitized.slice(0, -1);
-				if (potentialSingular.length > 1) return potentialSingular;
-			}
-			return sanitized;
-		}
-		
-		// Fallback naming if no other inference is possible
-		return 'entity_' + Math.random().toString(36).substr(2, 9);
-	}
-	
-	private recordRelationship(fromTable: string, toTable: string): void {
-		if (fromTable === toTable) return; // Avoid self-relationships
-		
-		const relationshipKey = `${fromTable}_${toTable}`;
-		const reverseKey = `${toTable}_${fromTable}`;
-		
-		const fromRelationships = this.entityRelationships.get(fromTable) || new Set();
-		const toRelationships = this.entityRelationships.get(toTable) || new Set();
-		
-		// Only record if not already recorded in either direction
-		if (!fromRelationships.has(toTable) && !toRelationships.has(fromTable)) {
-			fromRelationships.add(toTable);
-			this.entityRelationships.set(fromTable, fromRelationships);
-		}
-	}
-	
-	private createSchemasFromEntities(schemas: Record<string, TableSchema>): void {
-		// Create main entity tables
-		for (const [entityType, entities] of this.discoveredEntities.entries()) {
-			if (entities.length === 0) continue;
-			
-			const columnTypes: Record<string, Set<string>> = {};
-			const sampleData: any[] = [];
-			
-			entities.forEach((entity, index) => {
-				if (index < 3) {
-					sampleData.push(this.extractEntityFields(entity, columnTypes, entityType));
-				} else {
-					this.extractEntityFields(entity, columnTypes, entityType);
-				}
-			});
-			
-			const columns = this.resolveColumnTypes(columnTypes);
-			this.ensureIdColumn(columns);
-			
-			schemas[entityType] = {
-				columns,
-				sample_data: sampleData
-			};
-		}
-		
-		// Create junction tables for many-to-many relationships
-		this.createJunctionTableSchemas(schemas);
-	}
-	
-	private extractEntityFields(obj: any, columnTypes: Record<string, Set<string>>, entityType: string): any {
-		const rowData: any = {};
-		
-		if (!obj || typeof obj !== 'object') {
-			this.addColumnType(columnTypes, 'value', this.getSQLiteType(obj));
-			return { value: obj };
-		}
-		
-		for (const [key, value] of Object.entries(obj)) {
-			const columnName = this.sanitizeColumnName(key);
-			
-			if (Array.isArray(value)) {
-				// Check if this array contains entities that should be related
-				if (value.length > 0 && this.isEntity(value[0])) {
-					// This will be handled as a relationship via junction table, skip for now
-					continue;
-				} else {
-					// Store as JSON for analysis
-					this.addColumnType(columnTypes, columnName, 'JSON');
-					rowData[columnName] = JSON.stringify(value);
-				}
-			} else if (value && typeof value === 'object') {
-				if (this.isEntity(value)) {
-					// This is a related entity - create foreign key
-					const foreignKeyColumn = columnName + '_id';
-					this.addColumnType(columnTypes, foreignKeyColumn, 'INTEGER');
-					rowData[foreignKeyColumn] = (value as any).id || null;
-				} else {
-					// Complex object that's not an entity
-					if (this.hasScalarFields(value)) {
-						// Flatten simple fields with prefixed names
-						for (const [subKey, subValue] of Object.entries(value)) {
-							if (!Array.isArray(subValue) && typeof subValue !== 'object') {
-								const prefixedColumn = columnName + '_' + this.sanitizeColumnName(subKey);
-								this.addColumnType(columnTypes, prefixedColumn, this.getSQLiteType(subValue));
-								rowData[prefixedColumn] = typeof subValue === 'boolean' ? (subValue ? 1 : 0) : subValue;
-							}
-						}
-					} else {
-						// Store complex object as JSON
-						this.addColumnType(columnTypes, columnName, 'JSON');
-						rowData[columnName] = JSON.stringify(value);
-					}
-				}
-			} else {
-				// Scalar values
-				this.addColumnType(columnTypes, columnName, this.getSQLiteType(value));
-				rowData[columnName] = typeof value === 'boolean' ? (value ? 1 : 0) : value;
-			}
-		}
-		
-		return rowData;
-	}
-	
-	private hasScalarFields(obj: any): boolean {
-		if (!obj || typeof obj !== 'object') return false;
-		return Object.values(obj).some(value => 
-			typeof value !== 'object' || value === null
-		);
-	}
-	
-	private createJunctionTableSchemas(schemas: Record<string, TableSchema>): void {
-		const junctionTables = new Set<string>();
-		
-		for (const [fromTable, relatedTables] of this.entityRelationships.entries()) {
-			for (const toTable of relatedTables) {
-				// Create a consistent junction table name (alphabetical order to avoid duplicates)
-				const junctionName = [fromTable, toTable].sort().join('_');
-				
-				if (!junctionTables.has(junctionName)) {
-					junctionTables.add(junctionName);
-					
-					schemas[junctionName] = {
-						columns: {
-							id: 'INTEGER PRIMARY KEY AUTOINCREMENT',
-							[`${fromTable}_id`]: 'INTEGER',
-							[`${toTable}_id`]: 'INTEGER'
-						},
-						sample_data: []
-					};
-				}
-			}
-		}
-	}
-	
-	private createSchemaFromPrimitiveOrSimpleArray(data: any, tableName: string): TableSchema {
-		const columnTypes: Record<string, Set<string>> = {};
-		const sampleData: any[] = [];
-		
-		if (Array.isArray(data)) {
-			data.slice(0,3).forEach(item => {
-				// This simple version just takes the value, assumes items are scalar or will be JSON stringified.
-				const row = this.extractSimpleFields(item, columnTypes);
-				sampleData.push(row);
-			});
-			if (data.length > 3) {
-				data.slice(3).forEach(item => this.extractSimpleFields(item, columnTypes));
-			}
-		} else { // Scalar data
-			const row = this.extractSimpleFields(data, columnTypes);
-			sampleData.push(row);
-		}
-		
-		const columns = this.resolveColumnTypes(columnTypes);
-		// No automatic 'id' for these simple tables unless the data happens to have one.
-		if (!Object.keys(columns).includes('id') && !Object.keys(columns).includes('value')) {
-			// If only one column and it is not named 'value', rename it to value for consistency
-			const colNames = Object.keys(columns);
-			if(colNames.length === 1 && colNames[0] !== 'value'){
-				columns['value'] = columns[colNames[0]];
-				delete columns[colNames[0]];
-				// also update sample data key
-				sampleData.forEach(s => { s['value'] = s[colNames[0]]; delete s[colNames[0]]; });
-			}
-		}
-		if (Object.keys(columns).length === 0 && data === null) { // handle null input
-		    columns['value'] = 'TEXT';
-		}
-
-		return { columns, sample_data: sampleData };
-	}
-
-	private createSchemaFromObject(obj: any, tableName: string): TableSchema {
-		const columnTypes: Record<string, Set<string>> = {};
-		const rowData = this.extractSimpleFields(obj, columnTypes);
-		const columns = this.resolveColumnTypes(columnTypes);
-		// No automatic 'id' for this type of table either.
-		return { columns, sample_data: [rowData] };
-	}
-
-	private extractSimpleFields(obj: any, columnTypes: Record<string, Set<string>>): any {
-		const rowData: any = {};
-		
-		if (obj === null || typeof obj !== 'object') {
-			this.addColumnType(columnTypes, 'value', this.getSQLiteType(obj));
-			return { value: obj };
-		}
-		
-		if (Array.isArray(obj)) { // Should not happen if called from createSchemaFromPrimitiveOrSimpleArray with array items
-			this.addColumnType(columnTypes, 'array_data_json', 'TEXT');
-			return { array_data_json: JSON.stringify(obj) };
-		}
-
-		for (const [key, value] of Object.entries(obj)) {
-			const columnName = this.sanitizeColumnName(key);
-			if (value === null || typeof value !== 'object') {
-				this.addColumnType(columnTypes, columnName, this.getSQLiteType(value));
-				rowData[columnName] = typeof value === 'boolean' ? (value ? 1 : 0) : value;
-			} else {
-				// For nested objects/arrays within this simple structure, store as JSON.
-				this.addColumnType(columnTypes, columnName + '_json', 'TEXT');
-				rowData[columnName + '_json'] = JSON.stringify(value);
-			}
-		}
-		return rowData;
-	}
-	
-	private addColumnType(columnTypes: Record<string, Set<string>>, column: string, type: string): void {
-		if (!columnTypes[column]) columnTypes[column] = new Set();
-		columnTypes[column].add(type);
-	}
-	
-	private resolveColumnTypes(columnTypes: Record<string, Set<string>>): Record<string, string> {
-		const columns: Record<string, string> = {};
-		
-		for (const [columnName, types] of Object.entries(columnTypes)) {
-			if (types.size === 1) {
-				columns[columnName] = Array.from(types)[0];
-			} else {
-				// Mixed types - prefer TEXT > REAL > INTEGER
-				columns[columnName] = types.has('TEXT') ? 'TEXT' : types.has('REAL') ? 'REAL' : 'INTEGER';
-			}
-		}
-		
-		return columns;
-	}
-	
-	private ensureIdColumn(columns: Record<string, string>): void {
-		if (!columns.id) {
-			columns.id = "INTEGER PRIMARY KEY AUTOINCREMENT";
-		} else if (columns.id === "INTEGER") {
-			columns.id = "INTEGER PRIMARY KEY";
-		}
-	}
-	
-	private getSQLiteType(value: any): string {
-		if (value === null || value === undefined) return "TEXT";
-		switch (typeof value) {
-			case 'number': return Number.isInteger(value) ? "INTEGER" : "REAL";
-			case 'boolean': return "INTEGER";
-			case 'string': return "TEXT";
-			default: return "TEXT";
-		}
-	}
-	
-	private sanitizeTableName(name: string): string {
-		if (!name || typeof name !== 'string') {
-			return 'table_' + Math.random().toString(36).substr(2, 9);
-		}
-		
-		let sanitized = name
-			.replace(/[^a-zA-Z0-9_]/g, '_')
-			.replace(/_{2,}/g, '_')  // Replace multiple underscores with single
-			.replace(/^_|_$/g, '')  // Remove leading/trailing underscores
-			.toLowerCase();
-		
-		// Ensure it doesn't start with a number
-		if (/^[0-9]/.test(sanitized)) {
-			sanitized = 'table_' + sanitized;
-		}
-		
-		// Ensure it's not empty and not a SQL keyword
-		if (!sanitized || sanitized.length === 0) {
-			sanitized = 'table_' + Math.random().toString(36).substr(2, 9);
-		}
-		
-		// Handle SQL reserved words
-		const reservedWords = ['table', 'index', 'view', 'column', 'primary', 'key', 'foreign', 'constraint'];
-		if (reservedWords.includes(sanitized)) {
-			sanitized = sanitized + '_table';
-		}
-		
-		return sanitized;
-	}
-	
-	private sanitizeColumnName(name: string): string {
-		if (!name || typeof name !== 'string') {
-			return 'column_' + Math.random().toString(36).substr(2, 9);
-		}
-		
-		// Convert camelCase to snake_case
-		let snakeCase = name
-			.replace(/([A-Z])/g, '_$1')
-			.toLowerCase()
-			.replace(/[^a-zA-Z0-9_]/g, '_')
-			.replace(/_{2,}/g, '_')  // Replace multiple underscores with single
-			.replace(/^_|_$/g, ''); // Remove leading/trailing underscores
-		
-		// Ensure it doesn't start with a number
-		if (/^[0-9]/.test(snakeCase)) {
-			snakeCase = 'col_' + snakeCase;
-		}
-		
-		// Ensure it's not empty
-		if (!snakeCase || snakeCase.length === 0) {
-			snakeCase = 'column_' + Math.random().toString(36).substr(2, 9);
-		}
-		
-		// Handle common genomics abbreviations properly
-		const genomicsTerms: Record<string, string> = {
-			'entrezid': 'entrez_id',
-			'displayname': 'display_name',
-			'varianttype': 'variant_type',
-			'evidencelevel': 'evidence_level',
-			'evidencetype': 'evidence_type',
-			'evidencedirection': 'evidence_direction',
-			'sourcetype': 'source_type',
-			'molecularprofile': 'molecular_profile',
-			'genomicchange': 'genomic_change'
-		};
-		
-		const result = genomicsTerms[snakeCase] || snakeCase;
-		
-		// Handle SQL reserved words
-		const reservedWords = ['table', 'index', 'view', 'column', 'primary', 'key', 'foreign', 'constraint', 'order', 'group', 'select', 'from', 'where'];
-		if (reservedWords.includes(result)) {
-			return result + '_col';
-		}
-		
-		return result;
-	}
-}
-
-// Enhanced data insertion engine with relational support
-class DataInsertionEngine {
-	private processedEntities: Map<string, Map<any, number>> = new Map();
-	private relationshipData: Map<string, Set<string>> = new Map(); // Track actual relationships found in data
-	
-	async insertData(data: any, schemas: Record<string, TableSchema>, sql: any): Promise<void> {
-		// Reset state for new insertion
-		this.processedEntities.clear();
-		this.relationshipData.clear();
-
-		const schemaNames = Object.keys(schemas);
-
-		// Check if this is one of the simple fallback schemas
-		if (schemaNames.length === 1 && (schemaNames[0] === 'scalar_data' || schemaNames[0] === 'array_data' || schemaNames[0] === 'root_object')) {
-			const tableName = schemaNames[0];
-			const schema = schemas[tableName];
-			if (tableName === 'scalar_data' || tableName === 'root_object') {
-				await this.insertSimpleRow(data, tableName, schema, sql);
-			} else { // array_data
-				if (Array.isArray(data)) {
-					for (const item of data) {
-						await this.insertSimpleRow(item, tableName, schema, sql);
-					}
-				} else {
-					await this.insertSimpleRow(data, tableName, schema, sql); 
-				}
-			}
-			return;
-		}
-
-		// Phase 1: Insert all entities first (to establish primary keys)
-		await this.insertAllEntities(data, schemas, sql);
-		
-		// Phase 2: Handle relationships via junction tables (only for tables with data)
-		await this.insertJunctionTableRecords(data, schemas, sql);
-	}
-
-	private async insertAllEntities(obj: any, schemas: Record<string, TableSchema>, sql: any, path: string[] = []): Promise<void> {
-		if (!obj || typeof obj !== 'object') return;
-		
-		// Handle arrays of entities
-		if (Array.isArray(obj)) {
-			for (const item of obj) {
-				await this.insertAllEntities(item, schemas, sql, path);
-			}
-			return;
-		}
-		
-		// Handle GraphQL edges pattern
-		if (obj.edges && Array.isArray(obj.edges)) {
-			const nodes = obj.edges.map((edge: any) => edge.node).filter(Boolean);
-			for (const node of nodes) {
-				await this.insertAllEntities(node, schemas, sql, path);
-			}
-			return;
-		}
-		
-		// Handle individual entities
-		if (this.isEntity(obj)) {
-			const entityType = this.inferEntityType(obj, path);
-			if (schemas[entityType]) {
-				await this.insertEntityRecord(obj, entityType, schemas[entityType], sql);
-				
-				// Process nested entities and record relationships
-				await this.processEntityRelationships(obj, entityType, schemas, sql, path);
-			}
-		}
-		
-		// Recursively explore nested objects
-		for (const [key, value] of Object.entries(obj)) {
-			await this.insertAllEntities(value, schemas, sql, [...path, key]);
-		}
-	}
-	
-	private async processEntityRelationships(entity: any, entityType: string, schemas: Record<string, TableSchema>, sql: any, path: string[]): Promise<void> {
-		for (const [key, value] of Object.entries(entity)) {
-			if (Array.isArray(value) && value.length > 0) {
-				// Check if this array contains entities
-				const firstItem = value.find(item => this.isEntity(item));
-				if (firstItem) {
-					const relatedEntityType = this.inferEntityType(firstItem, [key]);
-					
-					// Process all entities in this array and record relationships
-					for (const item of value) {
-						if (this.isEntity(item) && schemas[relatedEntityType]) {
-							await this.insertEntityRecord(item, relatedEntityType, schemas[relatedEntityType], sql);
-							
-							// Track this relationship for junction table creation
-							const relationshipKey = [entityType, relatedEntityType].sort().join('_');
-							const relationships = this.relationshipData.get(relationshipKey) || new Set();
-							const entityId = this.getEntityId(entity, entityType);
-							const relatedId = this.getEntityId(item, relatedEntityType);
-							
-							if (entityId && relatedId) {
-								relationships.add(`${entityId}_${relatedId}`);
-								this.relationshipData.set(relationshipKey, relationships);
-							}
-							
-							// Recursively process nested entities
-							await this.processEntityRelationships(item, relatedEntityType, schemas, sql, [...path, key]);
-						}
-					}
-				}
-			} else if (value && typeof value === 'object' && this.isEntity(value)) {
-				// Single related entity
-				const relatedEntityType = this.inferEntityType(value, [key]);
-				if (schemas[relatedEntityType]) {
-					await this.insertEntityRecord(value, relatedEntityType, schemas[relatedEntityType], sql);
-					await this.processEntityRelationships(value, relatedEntityType, schemas, sql, [...path, key]);
-				}
-			}
-		}
-	}
-	
-	private async insertEntityRecord(entity: any, tableName: string, schema: TableSchema, sql: any): Promise<number | null> {
-		// Check if this entity was already processed
-		const entityMap = this.processedEntities.get(tableName) || new Map();
-		if (entityMap.has(entity)) {
-			return entityMap.get(entity)!;
-		}
-		
-		const rowData = this.mapEntityToSchema(entity, schema);
-		if (Object.keys(rowData).length === 0) return null;
-		
-		const columns = Object.keys(rowData);
-		const placeholders = columns.map(() => '?').join(', ');
-		const values = Object.values(rowData);
-		
-		// Use INSERT OR IGNORE to handle potential duplicates
-		const insertSQL = `INSERT OR IGNORE INTO ${tableName} (${columns.join(', ')}) VALUES (${placeholders})`;
-		sql.exec(insertSQL, ...values);
-		
-		// Get the inserted or existing ID
-		let insertedId: number | null = null;
-		if (rowData.id) {
-			// If we have the ID in the data, use it
-			insertedId = rowData.id;
-		} else {
-			// Otherwise get the last inserted row ID
-			insertedId = sql.exec(`SELECT last_insert_rowid() as id`).one()?.id || null;
-		}
-		
-		// Track this entity
-		if (insertedId) {
-			entityMap.set(entity, insertedId);
-			this.processedEntities.set(tableName, entityMap);
-		}
-		
-		return insertedId;
-	}
-	
-	private async insertJunctionTableRecords(data: any, schemas: Record<string, TableSchema>, sql: any): Promise<void> {
-		// Only create junction table records for relationships that actually have data
-		for (const [relationshipKey, relationshipPairs] of this.relationshipData.entries()) {
-			if (schemas[relationshipKey]) {
-				const [table1, table2] = relationshipKey.split('_');
-				
-				for (const pairKey of relationshipPairs) {
-					const [id1, id2] = pairKey.split('_').map(Number);
-					
-					const insertSQL = `INSERT OR IGNORE INTO ${relationshipKey} (${table1}_id, ${table2}_id) VALUES (?, ?)`;
-					sql.exec(insertSQL, id1, id2);
-				}
-			}
-		}
-	}
-	
-	private getEntityId(entity: any, entityType: string): number | null {
-		const entityMap = this.processedEntities.get(entityType);
-		return entityMap?.get(entity) || null;
-	}
-	
-	private mapEntityToSchema(obj: any, schema: TableSchema): any {
-		const rowData: any = {};
-		
-		if (!obj || typeof obj !== 'object') {
-			if (schema.columns.value) rowData.value = obj;
-			return rowData;
-		}
-		
-		for (const columnName of Object.keys(schema.columns)) {
-			if (columnName === 'id' && schema.columns[columnName].includes('AUTOINCREMENT')) {
-				continue;
-			}
-			
-			let value = null;
-			
-			// Handle foreign key columns
-			if (columnName.endsWith('_id') && !columnName.includes('_json')) {
-				const baseKey = columnName.slice(0, -3);
-				const originalKey = this.findOriginalKey(obj, baseKey);
-				if (originalKey && obj[originalKey] && typeof obj[originalKey] === 'object') {
-					value = (obj[originalKey] as any).id || null;
-				}
-			}
-			// Handle prefixed columns (from nested scalar fields)
-			else if (columnName.includes('_') && !columnName.endsWith('_json')) {
-				const parts = columnName.split('_');
-				if (parts.length >= 2) {
-					const baseKey = parts[0];
-					const subKey = parts.slice(1).join('_');
-					const originalKey = this.findOriginalKey(obj, baseKey);
-					if (originalKey && obj[originalKey] && typeof obj[originalKey] === 'object') {
-						const nestedObj = obj[originalKey];
-						const originalSubKey = this.findOriginalKey(nestedObj, subKey);
-						if (originalSubKey && nestedObj[originalSubKey] !== undefined) {
-							value = nestedObj[originalSubKey];
-							if (typeof value === 'boolean') value = value ? 1 : 0;
-						}
-					}
-				}
-			}
-			// Handle JSON columns
-			else if (columnName.endsWith('_json')) {
-				const baseKey = columnName.slice(0, -5);
-				const originalKey = this.findOriginalKey(obj, baseKey);
-				if (originalKey && obj[originalKey] && typeof obj[originalKey] === 'object') {
-					value = JSON.stringify(obj[originalKey]);
-				}
-			}
-			// Handle regular columns
-			else {
-				const originalKey = this.findOriginalKey(obj, columnName);
-				if (originalKey && obj[originalKey] !== undefined) {
-					value = obj[originalKey];
-					if (typeof value === 'boolean') value = value ? 1 : 0;
-					
-					// Skip arrays of entities (they're handled via junction tables)
-					if (Array.isArray(value) && value.length > 0 && this.isEntity(value[0])) {
-						continue;
-					}
-				}
-			}
-			
-			if (value !== null && value !== undefined) {
-				rowData[columnName] = value;
-			}
-		}
-		
-		return rowData;
-	}
-	
-	// Entity detection and type inference (reuse logic from SchemaInferenceEngine)
-	private isEntity(obj: any): boolean {
-		if (!obj || typeof obj !== 'object' || Array.isArray(obj)) return false;
-		
-		const hasId = obj.id !== undefined || obj._id !== undefined;
-		const fieldCount = Object.keys(obj).length;
-		const hasMultipleFields = fieldCount >= 2;
-		
-		const hasEntityFields = obj.name !== undefined || obj.title !== undefined || 
-			obj.description !== undefined || obj.type !== undefined;
-		
-		return hasId || (hasMultipleFields && hasEntityFields);
-	}
-	
-	private inferEntityType(obj: any, path: string[]): string {
-		if (obj.__typename) return this.sanitizeTableName(obj.__typename);
-		if (obj.type && typeof obj.type === 'string') return this.sanitizeTableName(obj.type);
-		
-		if (path.length > 0) {
-			const lastPath = path[path.length - 1];
-			if (lastPath === 'edges' && path.length > 1) {
-				return this.sanitizeTableName(path[path.length - 2]);
-			}
-			if (lastPath.endsWith('s') && lastPath.length > 1) {
-				return this.sanitizeTableName(lastPath.slice(0, -1));
-			}
-			return this.sanitizeTableName(lastPath);
-		}
-		
-		return 'entity_' + Math.random().toString(36).substr(2, 9);
-	}
-	
-	private sanitizeTableName(name: string): string {
-		if (!name || typeof name !== 'string') {
-			return 'table_' + Math.random().toString(36).substr(2, 9);
-		}
-		
-		let sanitized = name
-			.replace(/[^a-zA-Z0-9_]/g, '_')
-			.replace(/_{2,}/g, '_')  // Replace multiple underscores with single
-			.replace(/^_|_$/g, '')  // Remove leading/trailing underscores
-			.toLowerCase();
-		
-		// Ensure it doesn't start with a number
-		if (/^[0-9]/.test(sanitized)) {
-			sanitized = 'table_' + sanitized;
-		}
-		
-		// Ensure it's not empty and not a SQL keyword
-		if (!sanitized || sanitized.length === 0) {
-			sanitized = 'table_' + Math.random().toString(36).substr(2, 9);
-		}
-		
-		// Handle SQL reserved words
-		const reservedWords = ['table', 'index', 'view', 'column', 'primary', 'key', 'foreign', 'constraint'];
-		if (reservedWords.includes(sanitized)) {
-			sanitized = sanitized + '_table';
-		}
-		
-		return sanitized;
-	}
-	
-	private findOriginalKey(obj: any, sanitizedKey: string): string | null {
-		const keys = Object.keys(obj);
-		
-		// Direct match
-		if (keys.includes(sanitizedKey)) return sanitizedKey;
-		
-		// Find key that sanitizes to the same value
-		return keys.find(key => 
-			this.sanitizeColumnName(key) === sanitizedKey
-		) || null;
-	}
-	
-	private sanitizeColumnName(name: string): string {
-		if (!name || typeof name !== 'string') {
-			return 'column_' + Math.random().toString(36).substr(2, 9);
-		}
-		
-		// Convert camelCase to snake_case
-		let snakeCase = name
-			.replace(/([A-Z])/g, '_$1')
-			.toLowerCase()
-			.replace(/[^a-zA-Z0-9_]/g, '_')
-			.replace(/_{2,}/g, '_')  // Replace multiple underscores with single
-			.replace(/^_|_$/g, ''); // Remove leading/trailing underscores
-		
-		// Ensure it doesn't start with a number
-		if (/^[0-9]/.test(snakeCase)) {
-			snakeCase = 'col_' + snakeCase;
-		}
-		
-		// Ensure it's not empty
-		if (!snakeCase || snakeCase.length === 0) {
-			snakeCase = 'column_' + Math.random().toString(36).substr(2, 9);
-		}
-		
-		// Handle common genomics abbreviations properly
-		const genomicsTerms: Record<string, string> = {
-			'entrezid': 'entrez_id',
-			'displayname': 'display_name',
-			'varianttype': 'variant_type',
-			'evidencelevel': 'evidence_level',
-			'evidencetype': 'evidence_type',
-			'evidencedirection': 'evidence_direction',
-			'sourcetype': 'source_type',
-			'molecularprofile': 'molecular_profile',
-			'genomicchange': 'genomic_change'
-		};
-		
-		const result = genomicsTerms[snakeCase] || snakeCase;
-		
-		// Handle SQL reserved words
-		const reservedWords = ['table', 'index', 'view', 'column', 'primary', 'key', 'foreign', 'constraint', 'order', 'group', 'select', 'from', 'where'];
-		if (reservedWords.includes(result)) {
-			return result + '_col';
-		}
-		
-		return result;
-	}
-
-	private async insertSimpleRow(obj: any, tableName: string, schema: TableSchema, sql: any): Promise<void> {
-		const rowData = this.mapObjectToSimpleSchema(obj, schema);
-		if (Object.keys(rowData).length === 0 && !(tableName === 'scalar_data' && obj === null)) return; // Allow inserting null for scalar_data
-
-		const columns = Object.keys(rowData);
-		const placeholders = columns.map(() => '?').join(', ');
-		const values = Object.values(rowData);
-
-		const insertSQL = `INSERT INTO ${tableName} (${columns.join(', ')}) VALUES (${placeholders})`;
-		sql.exec(insertSQL, ...values);
-	}
-
-	private mapObjectToSimpleSchema(obj: any, schema: TableSchema): any {
-		const rowData: any = {};
-
-		if (obj === null || typeof obj !== 'object') {
-			if (schema.columns.value) { // For scalar_data or array_data of primitives
-				rowData.value = obj;
-			} else if (Object.keys(schema.columns).length > 0) {
-				// This case should ideally not be hit if schema generation is right for primitives
-				// but as a fallback, if there's a column, try to put it there.
-				const firstCol = Object.keys(schema.columns)[0];
-				rowData[firstCol] = obj;
-			}
-			return rowData;
-		}
-
-		if (Array.isArray(obj)) { // For root_object schemas where a field might be an array
-			// This function (mapObjectToSimpleSchema) is for a single row. If an array needs to be a column, it should be JSON.
-			// This case likely means the schema is `root_object` and `obj` is one of its fields being mapped.
-			// The schema definition for `root_object` via `extractSimpleFields` handles JSON stringification.
-			// So, this specific path in mapObjectToSimpleSchema might be redundant if schema is well-defined.
-			// For safety, if a column expects `_json` for this array, it will be handled by the loop below.
-		}
-
-		for (const columnName of Object.keys(schema.columns)) {
-			let valueToInsert = undefined;
-			let originalKeyFound = false;
-
-			if (columnName.endsWith('_json')) {
-				const baseKey = columnName.slice(0, -5);
-				const originalKey = this.findOriginalKey(obj, baseKey);
-				if (originalKey && obj[originalKey] !== undefined) {
-					valueToInsert = JSON.stringify(obj[originalKey]);
-					originalKeyFound = true;
-				}
-			} else {
-				const originalKey = this.findOriginalKey(obj, columnName);
-				if (originalKey && obj[originalKey] !== undefined) {
-					const val = obj[originalKey];
-					if (typeof val === 'boolean') {
-						valueToInsert = val ? 1 : 0;
-					} else if (typeof val === 'object' && val !== null) {
-						// This should not happen if schema is from extractSimpleFields, which JSONifies nested objects.
-						// If it does, it implies a mismatch. For safety, try to JSON stringify.
-						valueToInsert = JSON.stringify(val);
-					} else {
-						valueToInsert = val;
-					}
-					originalKeyFound = true;
-				}
-			}
-
-			if (originalKeyFound && valueToInsert !== undefined) {
-				rowData[columnName] = valueToInsert;
-			} else if (obj.hasOwnProperty(columnName) && obj[columnName] !== undefined){ // Direct match as last resort
-				// This handles cases where sanitized names might not be used or `findOriginalKey` fails but direct prop exists
-				const val = obj[columnName];
-				if (typeof val === 'boolean') valueToInsert = val ? 1:0;
-				else if (typeof val === 'object' && val !== null) valueToInsert = JSON.stringify(val);
-				else valueToInsert = val;
-				rowData[columnName] = valueToInsert;
-			}
-		}
-		return rowData;
-	}
-}
-
-// Pagination analyzer - clean utility
-class PaginationAnalyzer {
-	
-	static extractInfo(data: any): PaginationInfo {
-		const result: PaginationInfo = {
-			hasNextPage: false,
-			hasPreviousPage: false,
-			currentCount: 0,
-			totalCount: null,
-			endCursor: null,
-			startCursor: null
-		};
-		
-		const pageInfo = this.findPageInfo(data);
-		if (pageInfo) {
-			Object.assign(result, {
-				hasNextPage: pageInfo.hasNextPage || false,
-				hasPreviousPage: pageInfo.hasPreviousPage || false,
-				endCursor: pageInfo.endCursor,
-				startCursor: pageInfo.startCursor
-			});
-		}
-		
-		result.totalCount = this.findTotalCount(data);
-		result.currentCount = this.countCurrentItems(data);
-		
-		if (result.hasNextPage) {
-			result.suggestion = `Use pagination to get more than ${result.currentCount} records. Add "pageInfo { hasNextPage endCursor }" to your query and use "after: \\"${result.endCursor}\\"" for next page.`;
-		}
-		
-		return result;
-	}
-	
-	private static findPageInfo(obj: any): any {
-		if (!obj || typeof obj !== 'object') return null;
-		if (obj.pageInfo && typeof obj.pageInfo === 'object') return obj.pageInfo;
-		
-		for (const value of Object.values(obj)) {
-			const found = this.findPageInfo(value);
-			if (found) return found;
-		}
-		return null;
-	}
-	
-	private static findTotalCount(obj: any): number | null {
-		if (!obj || typeof obj !== 'object') return null;
-		if (typeof obj.totalCount === 'number') return obj.totalCount;
-		
-		for (const value of Object.values(obj)) {
-			const found = this.findTotalCount(value);
-			if (found !== null) return found;
-		}
-		return null;
-	}
-	
-	private static countCurrentItems(obj: any): number {
-		// Count edges arrays first
-		const edgesArrays: any[][] = [];
-		this.findEdgesArrays(obj, edgesArrays);
-		
-		if (edgesArrays.length > 0) {
-			return edgesArrays.reduce((sum, edges) => sum + edges.length, 0);
-		}
-		
-		// Fallback to general array counting
-		return this.countArrayItems(obj);
-	}
-	
-	private static findEdgesArrays(obj: any, result: any[][]): void {
-		if (!obj || typeof obj !== 'object') return;
-		if (Array.isArray(obj.edges)) result.push(obj.edges);
-		
-		for (const value of Object.values(obj)) {
-			this.findEdgesArrays(value, result);
-		}
-	}
-	
-	private static countArrayItems(obj: any): number {
-		if (!obj || typeof obj !== 'object') return 0;
-		
-		let count = 0;
-		for (const value of Object.values(obj)) {
-			if (Array.isArray(value)) {
-				count += value.length;
-			} else if (typeof value === 'object') {
-				count += this.countArrayItems(value);
-			}
-		}
-		return count;
-	}
-}
 
 // Main Durable Object class - clean and focused
 export class JsonToSqlDO extends DurableObject {
